update context: * data (temp and default) * search (removed) * loadState * api (get users) * removed storage (component)
