+++ src/components/context/context.js
@@ -1,33 +1,33 @@
-import React, { useState, createContext } from 'react';
-import { templateData } from '../../modules/templatedata';
+import React, { useState, createContext, useEffect } from 'react';
+import { getUsers } from '../../api/actions';
 
-export const DataManagerContext = createContext();
+export const DataStorageContext = createContext();
 
 // eslint-disable-next-line react/prop-types
-export const DataManagerContextProvider = ({ children }) => {
-  const [data, setData] = useState(templateData);
+export const DataStorageContextProvider = ({ children }) => {
+  const [data, setData] = useState([]);
+  const [defaultData, setDefaultData] = useState([]);
 
-  // setter and getter
+  const [isLoading, setLoading] = useState(true);
+
   const updateData = (dump) => setData(dump);
+  const initData = (dump) => {
+    setData(dump);
+    setDefaultData(dump);
+  }
 
-  const value = { data, updateData };
+  useEffect(() => {
+    getUsers(initData); // save fetch result into data (temp) and defaultData
+    setLoading(false); // hide load state
+  }, []);
+
+  const value = { data, updateData, isLoading, defaultData, setDefaultData };
 
   return (
-    <DataManagerContext.Provider value={value}>
+    <DataStorageContext.Provider value={value}>
       {children}
-    </DataManagerContext.Provider>
+    </DataStorageContext.Provider>
   );
 };
 
 export const SearchContext = createContext('');
-
-// eslint-disable-next-line react/prop-types
-export const SearchContextProvider = ({ children }) => {
-  const [search, setSearch] = useState('');
-
-  const value = { search, setSearch };
-
-  return (
-    <SearchContext.Provider value={value}>{children}</SearchContext.Provider>
-  );
-};
